5-QA Engine/requirements.txt
fastapi==0.66.0
farm==0.8.0
farm-haystack==0.9.0
uvicorn==0.14.0
torch==1.8.1
torchvision==0.9.1
torchaudio==0.8.1
pydantic==1.8.2